test: move hosted omit-mode payload tests under first-party line limit
Relocate rendered-stack credential omit regressions into owned part_001 so
test_hosted_delegation_payloads.py stays under the 1500-line gate.
